feat: update the sample app to test the screen navigation
